tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/toBlob.png.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: toBlob.png</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>toBlob.png</h1>
<p class="desc">toBlob with image/png returns a PNG Blob</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toBlob with image/png returns a PNG Blob");
_addTest(function(canvas, ctx) {

canvas.toBlob(function(data){
  _assertSame(data.type, "image/png", "data.type", "\"image/png\"");
}, 'image/png');


});
</script>
